requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.0
PyNaCl==1.5.0
youtube-dl==2021.12.17 